chore(conf): move configuration loading from `conf` to `confload`
The first step was moving configuration loading out of the conf
package to make it easier to reason about and create a clear api
boundary between loading and all other operations. This required
updating a good number of files, but no behavioral changes were
introduced so it is simple enough to audit.

The only other notable change was moving a couple e2ecfg functions
into a separate package to allow them to be used without causing
import cycles.
