+++ internal/api/oauthserver/handlers_test.go
@@ -14,6 +14,7 @@ import (
 	"github.com/stretchr/testify/suite"
 	"github.com/supabase/auth/internal/api/shared"
 	"github.com/supabase/auth/internal/conf"
+	"github.com/supabase/auth/internal/conf/confload"
 	"github.com/supabase/auth/internal/hooks/v0hooks"
 	"github.com/supabase/auth/internal/models"
 	"github.com/supabase/auth/internal/storage"
@@ -31,7 +32,7 @@ type OAuthClientTestSuite struct {
 }
 
 func TestOAuthClientHandler(t *testing.T) {
-	globalConfig, err := conf.LoadGlobal(oauthServerTestConfig)
+	globalConfig, err := confload.LoadGlobal(oauthServerTestConfig)
 	require.NoError(t, err)
 
 	conn, err := test.SetupDBConnection(globalConfig)
