+++ internal/api/oauthserver/service_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 	"github.com/supabase/auth/internal/conf"
+	"github.com/supabase/auth/internal/conf/confload"
 	"github.com/supabase/auth/internal/hooks/v0hooks"
 	"github.com/supabase/auth/internal/models"
 	"github.com/supabase/auth/internal/storage"
@@ -26,7 +27,7 @@ type OAuthServiceTestSuite struct {
 }
 
 func TestOAuthService(t *testing.T) {
-	globalConfig, err := conf.LoadGlobal(serviceTestConfig)
+	globalConfig, err := confload.LoadGlobal(serviceTestConfig)
 	require.NoError(t, err)
 
 	conn, err := test.SetupDBConnection(globalConfig)
